feature/api create project api first wizard hasd done and services based code and migration files are added

+++ database/migrations/2021_08_22_145214_create_user_preferences_table.php
@@ -15,32 +15,33 @@ class CreateUserPreferencesTable extends Migration
      */
     public function up()
     {
-        Schema::create('user_preferences', function (Blueprint $table) {
-            $table->id();
-
-            $table->foreignId('user_id');
-
-            $table->string('project_name')->index();
-
-            $table->foreignId('state_id');
-
-            $table->foreignId('project_type_id');
-
-            $table->foreignId('role_id');
-
-            $table->foreignId('customer_id');
-
-            $table->string('answer1')->nullable();
-
-            $table->string('answer2')->nullable();
-
-            $table->timestamps();
-            $table->softDeletes();
-            $table->unsignedBigInteger('created_by')->nullable();
-
-            $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
-
-        });
+        if (!Schema::hasTable('user_preferences')) {
+            Schema::create('user_preferences', function (Blueprint $table) {
+                $table->id();
+    
+                $table->foreignId('user_id');
+    
+                $table->string('project_name')->index();
+    
+                $table->foreignId('state_id');
+    
+                $table->foreignId('project_type_id');
+    
+                $table->foreignId('role_id');
+    
+                $table->foreignId('customer_id');
+    
+                $table->string('answer1')->nullable();
+    
+                $table->string('answer2')->nullable();
+    
+                $table->timestamps();
+                $table->softDeletes();
+                $table->unsignedBigInteger('created_by')->nullable();
+    
+                $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     /**
